mh2lm: add a shim for fingerprint HAL for RBS-based fingerprint devices
libRbsFlow hardcodes its userdata path to `/data/fpdata`, which
violates Treble compliance.

There are a few ways it detects which path to use:
- hardcoded directory path
- `g_custom_ini_path` symbol

Unfortunately, `g_custom_ini_path` is in BSS, so we cannot simply hack
it in. There are no blobs that externally modify this.

Fortunately, `g_custom_ini_path` is an exported string, so:
- Replace libc's `dlopen`
- When `dlopen` is called, and the library is not `libRbsFlow.so`,
  return the dlopen handle with no edits
- If it is `libRbsFlow.so`, modify `g_custom_ini_path` to a non-Treble
  violating path, create the directory, and return the dlopen handle

Co-authored-by: Bailey Fox <bfox200012@gmail.com>
Change-Id: If708db14176e0945a8c07a4735260a71c95fb779