timelm: commonize ro.vendor.lge.capp_cupss.rootdir [2/2]
This prop is checked by the audio HAL to load ACDBs from the /OP
partition as a fallback, and defines the location.

This is usually in /product/OP, but since using /product hits a
neverallow, we switched the mount point to /vendor/OP, so point it
there.

All devices need to point to the same place, so move the prop to
common.

Change-Id: I205dc08c836d8718751cafccfc14c3a3cb99cd41